# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Platform API",
        "description": "Multi-tenant event platform API.\n\n- **Global routes** (`/api/...`) — authentication and platform (super admin) endpoints.\n- **Tenant routes** (`/api/tenant/...`) — resolved via the venue subdomain; every endpoint\n  below this prefix is scoped to the authenticated user's venue.\n\nAll mutating endpoints follow the `FormRequest → DTO → Service → ApiResource → ApiResponse`\npipeline. **OpenAPI schemas in `app/OpenApi/` are a read-only projection** of that contract —\nFormRequest + ApiResource + DTO remain the source of truth; update both in the same commit.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Global platform API"
        },
        {
            "url": "/api/tenant",
            "description": "Tenant-scoped API (requires venue subdomain resolution)"
        }
    ],
    "paths": {
...
        "/api/tenant/organizer/settlement/summary": {
            "get": {
                "tags": [
                    "Settlement"
                ],
                "summary": "Organizer settlement summary",
                "operationId": "tenant.organizer.settlement.summary",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Settlement summary",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "summary": {
                                                    "$ref": "#/components/schemas/SettlementSummary"
                                                },
                                                "entries": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/SettlementLedgerEntry"
                                                    }
                                                },
                                                "payments": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "meta": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tenant/organizer/settlement/entries": {
            "get": {
                "tags": [
                    "Settlement"
                ],
                "summary": "Organizer settlement ledger entries",
                "operationId": "tenant.organizer.settlement.entries",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Settlement ledger entries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "summary": {
                                                    "$ref": "#/components/schemas/SettlementSummary"
                                                },
                                                "entries": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/SettlementLedgerEntry"
                                                    }
                                                },
                                                "payments": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "meta": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "description": "Generic API error envelope returned by `ApiExceptionRenderer`.",
                        "type": "string",
                        "example": "Resource not found."
                    }
                },
                "type": "object"
            },
...
            "SettlementSummary": {
                "properties": {
                    "gross_sales": {
                        "type": "string",
                        "example": "125000.00"
                    },
                    "tickets_sold": {
                        "type": "integer",
                        "example": 320
                    },
                    "commission_due": {
                        "type": "string",
                        "example": "1250.00"
                    },
                    "commission_paid": {
                        "type": "string",
                        "example": "500.00"
                    },
                    "commission_adjustments": {
                        "type": "string",
                        "example": "50.00"
                    },
                    "commission_outstanding": {
                        "type": "string",
                        "example": "700.00"
                    },
                    "refunds": {
                        "type": "string",
                        "example": "5000.00"
                    },
                    "net_sales": {
                        "type": "string",
                        "example": "120000.00"
                    },
                    "currency": {
                        "type": "string",
                        "example": "USD"
                    }
                },
                "type": "object"
            },
            "SettlementLedgerEntry": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "type": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "credit": {
                        "type": "string",
                        "example": "10.00"
                    },
                    "debit": {
                        "type": "string",
                        "example": "0.00"
                    },
                    "balance": {
                        "type": "string",
                        "example": "150.00"
                    },
                    "order_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "event_id": {
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum Personal Access Token",
                "description": "Laravel Sanctum bearer token. Obtain via /api/auth/login or /api/tenant/auth/login."
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}